+++ NEWS.md
@@ -1,5 +1,7 @@
 # ps (development version)
 
+* New `ps_string()` for uniquely identifying a process.
+
 # ps 1.9.1
 
 * ps now builds correctly on Alpine Linux (3.19) on R 4.5.0.